[MOD/#3] ConcatAdapter 적용
