+++ app/src/main/java/org/android/go/sopt/presentation/main/home/HomeFragment.kt
@@ -3,6 +3,7 @@ package org.android.go.sopt.presentation.main.home
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.ConcatAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import org.android.go.sopt.R
 import org.android.go.sopt.databinding.FragmentHomeBinding
@@ -12,7 +13,8 @@ import org.android.go.sopt.util.binding.BindingFragment
 class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home) {
     private val viewModel by viewModels<HomeViewModel>()
 
-    private val repoAdapter by lazy { RepoAdapter(requireContext()) }
+    private val repoHeaderAdapter by lazy { RepoHeaderAdapter() }
+    private val repoItemAdapter by lazy { RepoItemAdapter() }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -21,8 +23,8 @@ class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home
     }
 
     private fun initRepoListAdapter() {
-        repoAdapter.submitList(viewModel.mockRepoList)
-        binding.rvHomeRepo.adapter = repoAdapter
+        repoItemAdapter.submitList(viewModel.mockRepoList)
+        binding.rvHomeRepo.adapter = ConcatAdapter(repoHeaderAdapter, repoItemAdapter)
     }
 
     companion object {
